chore: add zone cli param and adjust default username and host
The icinga master node (docker) per default uses the zone `master` and username `root` as well as `localhost` as hostname. 

The exploit script tries to overwrite the config with the same username and hostname. This won't work targeting master nodes and will fail silently. The new config will be ignored and exploit fails. Can be simply fixed by using a custom username and hostname.

Moreover, the default zone `master` is not always the same to the node name, given in `--node-cn`. This will brick the exploit script as well. With this commit, we introduce a new `--zone` cli parameter. Uses the default `master` zone or the custom one given by the end user.